# packages/shared/openapi/openapi-generated.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kyro API",
    "version": "1.0.0",
    "description": "Language-agnostic API platform for file storage and management.",
    "contact": {
      "name": "Kyro Support",
      "url": "https://kyro.io"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Development"
    },
    {
      "url": "https://api.kyro.io/api/v1",
      "description": "Production"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string"
              },
              "code": {
                "type": "string"
              },
              "details": {
                "type": "array",
                "items": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
...
      "ApiKey": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "key_prefix": {
            "type": "string"
          },
          "scopes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "last_used_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "revoked_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid request body",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Authentication required",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Insufficient permissions",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "PayloadTooLarge": {
        "description": "File too large",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/files": {
      "get": {
        "tags": [
          "Files"
        ],
        "summary": "List files",
        "security": [
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 100,
              "maximum": 100
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "x-scope": "read",
        "x-body-description": null,
        "x-response-example": "[{\"id\":\"550e8400-e29b-41d4-a716-446655440000\",\"name\":\"document.pdf\",\"mimeType\":\"application/pdf\",\"sizeBytes\":1024000,\"createdAt\":\"2026-04-10T12:00:00Z\"}]",
        "responses": {
          "200": {
            "description": "List of files"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      },
      "post": {
        "tags": [
          "Files"
        ],
        "summary": "Upload file",
        "security": [
          {
            "apiKey": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "File to upload"
                  }
                }
              }
            }
          }
        },
        "x-scope": "write",
        "x-body-description": "file — File (multipart/form-data)",
        "x-response-example": "{\"id\":\"550e8400-e29b-41d4-a716-446655440000\",\"name\":\"document.pdf\",\"mimeType\":\"application/pdf\",\"sizeBytes\":1024000,\"createdAt\":\"2026-04-10T12:00:00Z\"}",
        "responses": {
          "201": {
            "description": "File uploaded"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          }
        }
      }
    },
    "/files/{id}": {
      "get": {
        "tags": [
          "Files"
        ],
        "summary": "Download file",
        "security": [
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "x-scope": "read",
        "x-body-description": null,
        "x-response-example": "Binary file data with appropriate Content-Type header.",
        "responses": {
          "200": {
            "description": "File content",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "tags": [
          "Files"
        ],
        "summary": "Delete file",
        "security": [
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "File UUID"
          }
        ],
        "x-scope": "write",
        "x-body-description": null,
        "x-response-example": "{\"message\":\"File deleted\",\"id\":\"550e8400-e29b-41d4-a716-446655440000\"}",
        "responses": {
          "200": {
            "description": "File deleted"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/keys": {
      "get": {
        "tags": [
          "API Keys"
        ],
        "summary": "List API keys",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-scope": null,
        "x-body-description": null,
        "x-response-example": "[{\"id\":\"550e8400-e29b-41d4-a716-446655440000\",\"name\":\"Production Key\",\"key_prefix\":\"kyr_abc123\",\"scopes\":[\"read\",\"write\"],\"last_used_at\":\"2026-04-10T12:00:00Z\",\"revoked_at\":null,\"created_at\":\"2026-04-10T12:00:00Z\"}]",
        "responses": {
          "200": {
            "description": "List of API keys",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApiKey"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "post": {
        "tags": [
          "API Keys"
        ],
        "summary": "Create API key",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 255,
                    "description": "Key name"
                  },
                  "scopes": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "enum": [
                        "read",
                        "write",
                        "admin"
                      ]
                    },
                    "default": [
                      "read"
                    ],
                    "description": "Key permissions"
                  }
                }
              }
            }
          }
        },
        "x-scope": null,
        "x-body-description": "{ \"name\": \"Production Key\", \"scopes\": [\"read\", \"write\"] }",
        "x-response-example": "{\"id\":\"550e8400-e29b-41d4-a716-446655440000\",\"name\":\"Production Key\",\"key_prefix\":\"kyr_abc123\",\"scopes\":[\"read\",\"write\"],\"key\":\"kyr_abc123xyz...\",\"created_at\":\"2026-04-10T12:00:00Z\"}",
        "responses": {
          "201": {
            "description": "API key created"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/keys/{id}": {
      "delete": {
        "tags": [
          "API Keys"
        ],
        "summary": "Revoke API key",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "x-scope": null,
        "x-body-description": null,
        "x-response-example": "{\"message\":\"API key revoked\",\"id\":\"550e8400-e29b-41d4-a716-446655440000\"}",
        "responses": {
          "200": {
            "description": "Key revoked"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  }
}